requirements.txt
flask==3.1.2
psycopg2==2.9.11
flask-sqlalchemy==3.1.1
flask_migrate==4.1.0
flask_login==0.6.3
flask_session==0.8.0
python-dotenv==1.1.1
argon2-cffi==25.1.0
cryptography==46.0.2
psycopg2-binary==2.9.11